requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
Faker==19.1.0
filelock==3.12.2
FoxPuppet==1.0.3
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsoncodable==0.1.7
jsonpickle==3.0.1
k-selenium-cookies==0.0.4
kproxy==0.0.1
noraise==0.0.16
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
PySocks==1.7.1
pytest==7.4.0
pytest-firefox==0.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
selenium==4.10.0
selenium-browser==0.0.15
selenium-firefox==2.0.8
selenium-webdriver-extender==0.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==3.4.4
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
xpath-utils==0.0.3
